Refactor build script data
