+++ crates/project_model/src/lib.rs
@@ -6,6 +6,7 @@ mod project_json;
 mod sysroot;
 mod workspace;
 mod rustc_cfg;
+mod build_data;
 
 use std::{
     fs::{read_dir, ReadDir},
